changes to order and cart

+++ src/api/v1/controllers/orderController.ts
@@ -7,7 +7,7 @@ import Product from "../models/Product";
 export const orderFetchController = async (req: Request, res: Response) => {
     // const orders = [
     //     {
-    //         orderId: "12345",
+    //         orderId: "83a4d796-7ae2-4469-9bc6-595ed7937349",
     //         item: {
     //             id: 1,
     //             imageUrl: "assets/images/petrol.jpg",
@@ -19,18 +19,13 @@ export const orderFetchController = async (req: Request, res: Response) => {
     //             quantity: "2",
     //         },
     //         status: "PENDING",
-    //         deliveryPartner: {
-    //             deliveryPartnerId: "DP123",
-    //             deliveryTime: "2-3 days",
-    //             deliveryPartnerMobileNumber: "9876543210",
-    //             deliveryPartnerLocation: {
-    //                 latitude: 88.8,
-    //                 longitude: 88.9,
-    //             },
+    //         deliveryLocation: {
+    //             latitude: 8.568625,
+    //             longitude: 76.9619567,
     //         },
     //     },
     //     {
-    //         orderId: "98753",
+    //         orderId: "83a4d796-9bc6-595ed7937349-7ae2-4469",
     //         item: {
     //             id: 1,
     //             imageUrl: "assets/images/petrol.jpg",
@@ -41,7 +36,7 @@ export const orderFetchController = async (req: Request, res: Response) => {
     //             value: 2,
     //             quantity: "2",
     //         },
-    //         status: "PENDING",
+    //         status: "COMPLETED",
     //         deliveryPartner: {
     //             deliveryPartnerId: "DP123",
     //             deliveryTime: "2-3 days",
@@ -51,24 +46,56 @@ export const orderFetchController = async (req: Request, res: Response) => {
     //                 longitude: 88.9,
     //             },
     //         },
+    //         deliveryLocation: {
+    //             latitude: 8.568625,
+    //             longitude: 76.9619567,
+    //         },
+    //     },
+    //     {
+    //         orderId: "83a4d796-9bc6-595ed7937349-7ae2-4469",
+    //         item: {
+    //             id: 1,
+    //             imageUrl: "assets/images/petrol.jpg",
+    //             name: "Product Name",
+    //             price: 9.99,
+    //             review: 4.5,
+    //             star: 4.0,
+    //             value: 2,
+    //             quantity: "2",
+    //         },
+    //         status: "DELIVERY",
+    //         deliveryPartner: {
+    //             deliveryPartnerId: "DP123",
+    //             deliveryTime: "2-3 days",
+    //             deliveryPartnerMobileNumber: "9876543210",
+    //             deliveryPartnerLocation: {
+    //                 latitude: 88.8,
+    //                 longitude: 88.9,
+    //             },
+    //         },
+    //         deliveryLocation: {
+    //             latitude: 8.568625,
+    //             longitude: 76.9619567,
+    //         },
     //     },
     // ];
+    const { userId } = req.query;
+console.log(userId);
 
     console.log("hit order");
-    const orders = await Order.find();
+    const orders = await Order.find({ userId: userId });
     res.status(200).json(orders);
 };
 
 export const orderGenerateController = async (req: Request, res: Response) => {
-    const { cartData, position } = req.body;
+    const { cartData, position, userId } = req.body;
 
     console.log("hit");
 
     try {
         cartData.map((data: any) => {
-            console.log(data.cartId);
-
             const orderItem = new Order({
+                userId: userId,
                 orderId: uuidv4(),
                 item: new Product({
                     ...data.item,
@@ -79,7 +106,7 @@ export const orderGenerateController = async (req: Request, res: Response) => {
             orderItem.save();
         });
 
-        await Cart.deleteMany();
+        await Cart.deleteMany({ userId: userId });
         res.status(200).json({
             status: "success",
             message: "Order placed successfully.",
